Reduce non-determinism in monomorphiser
We were picking the first type we see which satisfies any necessary constraints,
but this is non-deterministic. Since we have a default on hand (to try in case
we can't find any), we now check to see if that's in the list of satisfying
types, and if so we use it.

This means that, for example: `Ord b => a -> b` with a default of `Integer` will
now monomorphise to `Integer -> Integer`, rather than `Integer -> X` where `X`
depends on the context (typeclass instances in scope, etc.)